quick-commerce-medicine-delivery-app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pillow
streamlit
requests
python-dateutil
typing-extensions
email-validator
geopy
folium
streamlit-folium